Job order

+++ src/com/smart/mis/client/function/production/process/ProcessData.java
@@ -5,7 +5,7 @@ import com.smartgwt.client.widgets.grid.ListGridRecord;
 public class ProcessData {
 
     //public static ListGridRecord createRecord(String id, String type, String desc, Double std_time, String pid, String mid, String mat_name) { 
-    public static ListGridRecord createRecord(String id, String type, String desc, Double std_time, String pid, Integer priority) {
+    public static ListGridRecord createRecord(String id, String type, String desc, Integer std_time, String pid, Integer priority) {
         ListGridRecord record = new ListGridRecord();
         record.setAttribute("pid", pid);
         record.setAttribute("psid", id);
@@ -18,7 +18,7 @@ public class ProcessData {
         return record;  
     }
     
-    public static ListGridRecord createRecord(String type, String desc, Double std_time, Integer priority) {
+    public static ListGridRecord createRecord(String type, String desc, Integer std_time, Integer priority) {
         ListGridRecord record = new ListGridRecord();
         record.setAttribute("type",type);  
         record.setAttribute("desc", desc);
@@ -34,17 +34,17 @@ public class ProcessData {
     
     if (pid != null && (pid.equals( "PD10001") || pid.equals( "PD10002") || pid.equals( "PD10003"))) {
     	return new ListGridRecord[]{ 
-    			createRecord("PS10001","1_casting", "", 3.0, "PD10001", 1),
-    			createRecord("PS10002","2_scrape", "", 3.0, "PD10001", 2),
-    			createRecord("PS10003","3_abrade", "", 3.0, "PD10001", 3),
-    			createRecord("PS10004","4_packing", "", 1.0, "PD10001", 4)
+    			createRecord("PS10001","1_casting", "", 3, "PD10001", 1),
+    			createRecord("PS10002","2_scrape", "", 3, "PD10001", 2),
+    			createRecord("PS10003","3_abrade", "", 3, "PD10001", 3),
+    			createRecord("PS10004","4_packing", "", 1, "PD10001", 4)
     	};
     } else if (pid != null && (pid.equals( "PD10004") || pid.equals( "PD10005") || pid.equals( "PD10006") || pid.equals( "PD10007"))) {
     	return new ListGridRecord[]{ 
-    			createRecord("PS20001","1_casting", "", 3.0, "PD10004", 1),
-    			createRecord("PS20002","2_scrape", "", 3.0, "PD10004", 2),
-    			createRecord("PS20003","3_abrade", "", 3.0, "PD10004", 3),
-    			createRecord("PS20004","4_packing", "", 1.0, "PD10004", 4)
+    			createRecord("PS20001","1_casting", "", 3, "PD10004", 1),
+    			createRecord("PS20002","2_scrape", "", 3, "PD10004", 2),
+    			createRecord("PS20003","3_abrade", "", 3, "PD10004", 3),
+    			createRecord("PS20004","4_packing", "", 1, "PD10004", 4)
     	};
     } 
     	else return new ListGridRecord[]{};
